+++ backend/packages/gates/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/gates
 
+## 0.5.39
+
+### Patch Changes
+
+- Updated dependencies [6108525]
+  - @cat-factory/kernel@0.129.1
+
 ## 0.5.38
 
 ### Patch Changes